Move ClowderController global into clowder_controller
