--- clowder/clowder/cli/globals.py
@@ -1,11 +1,0 @@
-# -*- coding: utf-8 -*-
-"""Clowder command line global variables
-
-.. codeauthor:: Joe Decapo <joe@polka.cat>
-
-"""
-
-from clowder.clowder_controller import ClowderController
-
-
-CLOWDER_CONTROLLER = ClowderController()
