+++ clowder/clowder/cli/sync_controller.py
@@ -12,7 +12,7 @@ import sys
 from cement.ext.ext_argparse import ArgparseController, expose
 from termcolor import cprint
 
-from clowder.cli.globals import CLOWDER_CONTROLLER
+from clowder.clowder_controller import CLOWDER_CONTROLLER
 from clowder.cli.parallel import sync
 from clowder.cli.util import options_help_message
 from clowder.clowder_repo import print_clowder_repo_status_fetch
