+++ examples/plugins/plugins/exampleplugin.py
@@ -9,11 +9,8 @@ from __future__ import print_function
 
 from cement.ext.ext_argparse import ArgparseController, expose
 
-from clowder.cli.globals import CLOWDER_CONTROLLER
-from clowder.clowder_repo import (
-    CLOWDER_REPO,
-    print_clowder_repo_status
-)
+from clowder.clowder_controller import CLOWDER_CONTROLLER
+from clowder.clowder_repo import print_clowder_repo_status
 import clowder.util.formatting as fmt
